requirements.txt
torch
torchvision
streamlit
numpy
pillow
opencv-python-headless
scikit-learn
grad-cam
reportlab